+++ src/Exception/FileNotFoundException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace KG\DigiDoc\Exception;
+
+/**
+ * @todo Have a custom exception message.
+ */
+class FileNotFoundException extends RuntimeException
+{
+
+}