+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 0.1.2 (unreleased)
  - Rename Container => Envelope
+ - Model OCSP Response and Request workflow
 
 0.1.1 2014-07-15